chore(suite-native): split exchange flow logic
